+++ Runtime/AbstractGameDataManager.cs
@@ -6,20 +6,24 @@ using ActionCode.Persistence;
 
 namespace ActionCode.GameDataSystem
 {
-    public abstract class AbstractGameDataSettings<T> : ScriptableObject, IDataModel where T : AbstractGameData
+    /// <summary>
+    /// Provides an abstract base class for managing game data, including local and cloud-based persistence.
+    /// </summary>
+    /// <remarks>
+    /// This class handles operations such as saving, loading, and deleting game data both locally
+    /// and remotely. It also provides support for cloud-based persistence if a cloud provider is available. Derived
+    /// classes can override certain methods to customize local persistence behavior.
+    /// </remarks>
+    /// <typeparam name="T">The type of game data managed by this class. Must derive from <see cref="AbstractGameData"/>.</typeparam>
+    public abstract class AbstractGameDataManager<T> : ScriptableObject, IGameDataManager where T : AbstractGameData
     {
         [SerializeField] private T gameData;
-        [SerializeField] private PersistenceSettings persistenceSettings;
+        [SerializeField] private PersistenceSettings persistence;
 
         /// <summary>
         /// The Game main data.
         /// </summary>
         public T Data => gameData;
-
-        /// <summary>
-        /// The serialization settings.
-        /// </summary>
-        public PersistenceSettings Serialization => persistenceSettings;
 
         public ICloudProvider CloudProvider => cloudProvider.Value;
 
@@ -30,7 +34,7 @@ namespace ActionCode.GameDataSystem
         public void UpdateData(object data)
         {
             var gameData = data as T;
-            var serializer = persistenceSettings.GetFileSystem().Serializer;
+            var serializer = persistence.GetFileSystem().Serializer;
             var content = serializer.Serialize(gameData);
 
             serializer.Deserialize(content, ref this.gameData);
@@ -44,20 +48,20 @@ namespace ActionCode.GameDataSystem
 
             if (HasCloudProvider())
             {
-                var name = persistenceSettings.GetSlotName(slot);
+                var name = persistence.GetSlotName(slot);
                 await CloudProvider.SaveAsync(Data, name);
             }
         }
 
-        public async Awaitable LoadFromLastSlotAsync() => await persistenceSettings.TryLoadLastSlot(Data);
-        public async Awaitable LoadLocallyAsync(int slot) => await persistenceSettings.TryLoad(Data, slot);
+        public async Awaitable LoadFromLastSlotAsync() => await persistence.TryLoadLastSlot(Data);
+        public async Awaitable LoadLocallyAsync(int slot) => await persistence.TryLoad(Data, slot);
 
         public async Awaitable LoadRemotelyAsync(int slot)
         {
-            var name = persistenceSettings.GetSlotName(slot);
+            var name = persistence.GetSlotName(slot);
             var json = await CloudProvider?.LoadAsync(name);
             var hasJson = !string.IsNullOrEmpty(json);
-            if (hasJson) persistenceSettings.GetFileSystem().Serializer.Deserialize(json, ref gameData);
+            if (hasJson) persistence.GetFileSystem().Serializer.Deserialize(json, ref gameData);
         }
 
         public async Awaitable<bool> IsContinueAvailable() => await HasLastSlotAvailable();
@@ -75,14 +79,14 @@ namespace ActionCode.GameDataSystem
 
         public async Awaitable<IList> ListSlotsAsync()
         {
-            var names = persistenceSettings.GetNames();
+            var names = persistence.GetNames();
             var slots = new T[names.Count()];
 
             for (var i = 0; i < slots.Length; i++)
             {
                 var slotName = names.ElementAt(i);
                 slots[i] = CreateInstance<T>();
-                await persistenceSettings.TryLoad(slots[i], slotName);
+                await persistence.TryLoad(slots[i], slotName);
             }
 
             return slots;
@@ -92,7 +96,7 @@ namespace ActionCode.GameDataSystem
         {
             var remoteData = await CloudProvider?.LoadAllAsync(playerId);
             var gamesData = new T[remoteData.Count()];
-            var serializer = persistenceSettings.GetFileSystem().Serializer;
+            var serializer = persistence.GetFileSystem().Serializer;
 
             for (var i = 0; i < gamesData.Length; i++)
             {
@@ -105,15 +109,15 @@ namespace ActionCode.GameDataSystem
         }
 
         protected virtual async Awaitable SaveLocallyAsync(ScriptableObject data, int slot) =>
-            await persistenceSettings.Save(data, slot);
+            await persistence.Save(data, slot);
 
-        protected virtual bool TryDeleteLocally(int slot) => persistenceSettings.TryDelete(slot);
-        protected virtual bool TryDeleteAllLocally() => persistenceSettings.TryDeleteAll();
+        protected virtual bool TryDeleteLocally(int slot) => persistence.TryDelete(slot);
+        protected virtual bool TryDeleteAllLocally() => persistence.TryDeleteAll();
 
         private async Awaitable<bool> HasLastSlotAvailable()
         {
             var data = CreateInstance<T>();
-            return await persistenceSettings.TryLoadLastSlot(data);
+            return await persistence.TryLoadLastSlot(data);
         }
 
         // The only available Cloud provider for now is from Unity Service.