+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- AbstractGameDataManager with custom Editor
+- IGameDataManager
 
-### Removed
-- Remove SerializableInterface dependency
 
 ## [0.1.0] - 2025-04-12
 ### Added